feat: Add `New-PasswordConfigurationExternalUsers` with Name parameter
The recent code changes introduce a new parameter `Name` to the `New-PasswordConfigurationExternalUsers` function. This parameter allows specifying the name of the configuration, which will be visible in HTML reports.

Based on the recent user commits and repository history, it seems that this change is part of adding support for other systems and making improvements to the module.

Commit message: feat: Add `New-PasswordConfigurationExternalUsers` with Name parameter